rearrange tracing-web, lint
