+++ packages/tracing-web/src/otel/index.ts
@@ -1,0 +1,1 @@
+export { GrafanaAgentTraceExporter } from './agent-exporter';
