+++ packages/web/src/globals.ts
@@ -2,6 +2,6 @@ import type { Agent } from '@grafana/agent-core';
 
 declare global {
   interface Window {
-    grafanaAgent: Agent
+    grafanaAgent: Agent;
   }
 }
